Added new tests, observer, shopkeeper functions

+++ src/serverConstants/ConfigConstants.java
@@ -1,10 +1,13 @@
 package serverConstants;
 
+/**
+ * Class for configuration constants used to preset variables
+ */
 public class ConfigConstants {
 	
 	private static Integer RetourePercentage = 10;
-	private static Integer PresetAccountBalance = 1000;
-	private static Integer PresetAccountLowerLimit = 0;
+	private static long PresetAccountBalance = 1000;
+	private static long PresetAccountLowerLimit = 0;
 
 	
 	private ConfigConstants() {}
@@ -26,28 +29,28 @@ public class ConfigConstants {
 	/**
 	 * @return the presetAccountBalance
 	 */
-	public static final Integer getPresetAccountBalance() {
+	public static final long getPresetAccountBalance() {
 		return PresetAccountBalance;
 	}
 
 	/**
 	 * @param presetAccountBalance the presetAccountBalance to set
 	 */
-	public static final void setPresetAccountBalance(Integer presetAccountBalance) {
+	public static final void setPresetAccountBalance(long presetAccountBalance) {
 		PresetAccountBalance = presetAccountBalance;
 	}
 
 	/**
 	 * @return the presetAccountLowerLimit
 	 */
-	public static final Integer getPresetAccountLowerLimit() {
+	public static final long getPresetAccountLowerLimit() {
 		return PresetAccountLowerLimit;
 	}
 
 	/**
 	 * @param presetAccountLowerLimit the presetAccountLowerLimit to set
 	 */
-	public static final void setPresetAccountLowerLimit(Integer presetAccountLowerLimit) {
+	public static final void setPresetAccountLowerLimit(long presetAccountLowerLimit) {
 		PresetAccountLowerLimit = presetAccountLowerLimit;
 	}
 }
